added rest service for calculating working days

+++ vacation-core/src/main/kotlin/org/voelk/vacation/Vacation.kt
@@ -13,7 +13,7 @@ class Vacation(private val start: LocalDate, private val end: LocalDate) {
             throw IllegalArgumentException("Maximum allowed period is 1000 days. Your period is $durationInDays days")
     }
 
-    fun durationInDays(): Long {
-        return ChronoUnit.DAYS.between(start, end.plusDays(1))
+    fun durationInDays(): Int {
+        return ChronoUnit.DAYS.between(start, end.plusDays(1)).toInt()
     }
 }
